moved testData.jsx to client/src, and updated references in App.jsx, RatingsReviews/Index.jsx, and Overview.jsx
